requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django>3.2,<3.3
django-autoslug==1.9.9
django-extensions==3.2.3
django-taggit==1.1.0
pillow==10.4.0
pytz==2024.1
sqlparse==0.5.0
typing-extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
